Improve dataclass handling
