+++ src/fairseq2/__init__.py
@@ -4,25 +4,22 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+
+import fairseq2n  # Report any fairseq2n initialization error eagerly.
+
+# isort: split
+
 from importlib_metadata import entry_points
+
+from fairseq2.datasets import _register_datasets
+from fairseq2.models import _register_models
 
 __version__ = "0.3.0.dev0"
 
 
-# Report any fairseq2n initialization error eagerly.
-import fairseq2n
-
-# isort: split
-
-from fairseq2.datasets import _register_datasets
-from fairseq2.models import _register_models
-from fairseq2.utils.yaml import _register_yaml_representers
-
-_register_datasets()
-
 _register_models()
 
-_register_yaml_representers()
+_register_datasets()
 
 
 def setup_extensions() -> None:
@@ -33,5 +30,5 @@ def setup_extensions() -> None:
             setup_extension()
         except TypeError:
             raise RuntimeError(
-                f"The entry point '{entry_point.value}' is not a valid fairseq2 extension setup function."
+                f"The entry point '{entry_point.value}' is not a valid fairseq2 setup function."
             )
